# services/mock-upstream/openapi.yaml
openapi: 3.1.0
info:
  title: 商情局 Mock Upstream API
  version: 1.0.0
  description: |
    面向联调与验收的合成数据 API。返回数据不代表真实企业、价格或上游能力。
    对下游契约保持稳定，真实上游到位后由 real-adapter 转换。
servers:
  - url: http://127.0.0.1:4190
    description: 本地 Mock
security:
  - ApiKey: []
paths:
...
  /open/v1/global/companies/search:
    get:
      summary: A01 全球数据库主体搜索
      description: 仅路由到全球企业数据库 Provider，不访问国内数据库。
      parameters:
        - { name: q, in: query, required: true, schema: { type: string }, example: Northstar }
        - { name: country, in: query, required: false, schema: { type: string, minLength: 2, maxLength: 2 }, example: US }
        - { $ref: "#/components/parameters/MockScenario" }
      responses:
        "200": { $ref: "#/components/responses/Envelope" }
        "400": { $ref: "#/components/responses/ErrorEnvelope" }
  /open/v1/domestic/companies/search:
    get:
      summary: A01 国内数据库主体搜索
      description: 仅路由到中国大陆企业数据库 Provider，不访问全球数据库。
      parameters:
        - { name: q, in: query, required: true, schema: { type: string }, example: 上海青岚科技 }
        - { $ref: "#/components/parameters/MockScenario" }
      responses:
        "200": { $ref: "#/components/responses/Envelope" }
        "400": { $ref: "#/components/responses/ErrorEnvelope" }
  /open/v1/companies/search:
    get:
      summary: A01 全球主体搜索与消歧（兼容旧路径）
      parameters:
        - { name: q, in: query, required: true, schema: { type: string }, example: Northstar }
        - { name: country, in: query, required: false, schema: { type: string, minLength: 2, maxLength: 2 }, example: US }
        - { $ref: "#/components/parameters/MockScenario" }
      responses:
        "200": { $ref: "#/components/responses/Envelope" }
        "400": { $ref: "#/components/responses/ErrorEnvelope" }
        "503": { $ref: "#/components/responses/ErrorEnvelope" }
...
  /open/v1/companies/{companyId}/modules/{moduleCode}:
    get:
      summary: A02 企业模块数据
      parameters:
        - { $ref: "#/components/parameters/CompanyId" }
        - name: moduleCode
          in: path
          required: true
          schema: { type: string, enum: [M01, M02, M03, M04, M05, M06, M07, M08, M09, M10] }
        - { $ref: "#/components/parameters/MockScenario" }
      responses:
        "200": { $ref: "#/components/responses/Envelope" }
        "422": { $ref: "#/components/responses/ErrorEnvelope" }
        "503": { $ref: "#/components/responses/ErrorEnvelope" }
...
components:
  securitySchemes:
    ApiKey: { type: apiKey, in: header, name: X-API-Key, description: 商情局对客统一鉴权凭证，仅通过 HTTPS 请求头传输；上游凭证由服务端 Provider Adapter 管理。 }
  parameters:
    CompanyId:
      name: companyId
      in: path
      required: true
      schema: { type: string }
      example: SQJ-DEMO-US-0001
    MockScenario:
      name: X-Mock-Scenario
      in: header
      required: false
      schema:
        type: string
        enum: [default, no-record, partial, ambiguous, no-coverage, provider-error, slow]
  schemas:
    DataState:
      type: string
      enum: [AVAILABLE, NO_RECORD, PARTIAL, NO_COVERAGE, AMBIGUOUS, PROVIDER_ERROR, SYSTEM_ERROR]
    Envelope:
      type: object
      required: [requestId, traceId, dataState, queryAt, cacheHit, billable, provider, data]
      properties:
        requestId: { type: string }
        traceId: { type: string }
        dataState: { $ref: "#/components/schemas/DataState" }
        queryAt: { type: string, format: date-time }
        sourceUpdatedAt: { type: string, format: date-time }
        cacheHit: { type: boolean }
        billable: { type: boolean }
        provider:
          type: object
          properties:
            mode: { type: string, enum: [MOCK, REAL] }
            code: { type: string }
        data: {}
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
  responses:
    Envelope:
      description: 统一响应
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
    ErrorEnvelope:
      description: 统一错误响应
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }